% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{as.numeric}
\alias{as.numeric}
\title{Coerce sparse_numeric to numeric}
\description{
Coerce sparse_numeric to numeric
}
\seealso{
Other coercion: 
\code{\link{as.sparse_numeric}}
}
\concept{coercion}
