% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{as.sparse_numeric}
\alias{as.sparse_numeric}
\title{Coerce numeric to sparse_numeric}
\description{
Coerce numeric to sparse_numeric
}
\seealso{
Other coercion: 
\code{\link{as.numeric}()}
}
\concept{coercion}
