% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_class.R
\name{coerce-sparse_numeric-numeric}
\alias{coerce-sparse_numeric-numeric}
\title{Coerce Sparse Numeric to Numeric Vector}
\arguments{
\item{from}{A \code{sparse_numeric} object}
}
\value{
A numeric vector
}
\description{
Converts a sparse_numeric object to a regular numeric vector by
filling in zeros at appropriate positions.
}
\examples{
x <- as(c(1, 0, 0, 4), "sparse_numeric")
y <- as(x, "numeric")

}
