% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_6.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Mean for sparse_numeric}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{sparse_numeric object}

\item{...}{additional arguments (ignored)}
}
\value{
numeric scalar
}
\description{
Computes the mean of a sparse_numeric vector including zeros.
}
