% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{norm,sparse_numeric,ANY-method}
\alias{norm,sparse_numeric,ANY-method}
\title{Norm of a sparse_numeric Object}
\usage{
\S4method{norm}{sparse_numeric,ANY}(x)
}
\arguments{
\item{x}{A sparse_numeric vector}
}
\description{
Operates on sparse vectors and computes the norm of the vector, which in this case is the square root of the sum of the squared individual elements of a vector.
}
