% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Euclidean norm of a sparse_numeric vector}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Unused.}
}
\value{
A numeric scalar.
}
\description{
Computes the Euclidean (L2) norm of the underlying full vector:
\deqn{ \|x\|_2 = \sqrt{\sum_i x_i^2} }
using only stored non-zero values.
}
