% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{numeric_to_sparse_numeric}
\alias{numeric_to_sparse_numeric}
\title{Coercion: numeric → sparse_numeric}
\arguments{
\item{from}{A numeric vector.}
}
\value{
A sparse_numeric object.
}
\description{
Converts a standard numeric vector into a sparse_numeric object.
}
