% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_6.R
\name{sort,sparse_numeric-method}
\alias{sort,sparse_numeric-method}
\title{Sort sparse_numeric by position}
\usage{
\S4method{sort}{sparse_numeric}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{sparse_numeric object}

\item{decreasing}{logical, whether to sort in decreasing order}

\item{...}{additional arguments (ignored)}
}
\value{
sparse_numeric object with sorted pos and value
}
\description{
Returns a sparse_numeric object with entries sorted by the `pos` slot.
}
