% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_6.R
\name{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{Add two sparse_numeric objects}
\usage{
\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{sparse_numeric object}

\item{y}{sparse_numeric object}
}
\value{
sparse_numeric object
}
\description{
Performs element-wise addition of two sparse_numeric objects.
}
\examples{
x <- as(c(0,2,0,3), "sparse_numeric")
y <- as(c(1,0,4,0), "sparse_numeric")
sparse_add(x, y)
}
