% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_6.R
\name{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Cross product of two sparse_numeric objects}
\usage{
\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{sparse_numeric object}

\item{y}{sparse_numeric object}
}
\value{
numeric scalar
}
\description{
Computes the dot product of two sparse_numeric objects.
}
\examples{
x <- as(c(1,2,0,3), "sparse_numeric")
y <- as(c(0,2,5,1), "sparse_numeric")
sparse_crossprod(x, y)
}
