% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric-method}
\title{Computes \eqn{\sum_i x_i y_i}.}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}.}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric scalar.
}
\description{
Computes \eqn{\sum_i x_i y_i}.
}
