% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_6.R
\name{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Multiply two sparse_numeric objects}
\usage{
\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{sparse_numeric object}

\item{y}{sparse_numeric object}
}
\value{
sparse_numeric object
}
\description{
Performs element-wise multiplication of two sparse_numeric objects.
}
\examples{
x <- as(c(0,2,0,3), "sparse_numeric")
y <- as(c(1,0,5,2), "sparse_numeric")
sparse_mult(x, y)
}
