% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_6.R
\name{sparse_norm,sparse_numeric-method}
\alias{sparse_norm,sparse_numeric-method}
\title{Euclidean norm of sparse_numeric object}
\usage{
\S4method{sparse_norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{sparse_numeric object}
}
\value{
numeric scalar
}
\description{
Computes the Euclidean norm of a sparse_numeric object.
}
\examples{
x <- as(c(3,0,4), "sparse_numeric")
sparse_norm(x)
}
