% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_6.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse numeric vector class}
\description{
This S4 class represents a numeric vector stored in sparse format,
storing only non-zero values and their positions.
}
\section{Slots}{

\describe{
\item{\code{value}}{Numeric vector of non-zero values.}

\item{\code{pos}}{Integer vector of positions corresponding to `value`.}

\item{\code{length}}{Integer of total vector length.}
}}

