% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{sparse_numeric_to_numeric}
\alias{sparse_numeric_to_numeric}
\title{Coercion: sparse_numeric → numeric}
\arguments{
\item{from}{A sparse_numeric object.}
}
\value{
A numeric vector.
}
\description{
Converts a sparse_numeric object back into a full numeric vector.
}
