% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_6.R
\name{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtract two sparse_numeric objects}
\usage{
\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{sparse_numeric object}

\item{y}{sparse_numeric object}
}
\value{
sparse_numeric object
}
\description{
Performs element-wise subtraction of y from x.
}
\examples{
x <- as(c(1,2,0,3), "sparse_numeric")
y <- as(c(1,1,0,4), "sparse_numeric")
sparse_sub(x, y)
}
