% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\title{Subtract two sparse_numeric vectors}
\usage{
sparse_sub(x, y, ...)

\S4method{sparse_sub}{sparse_numeric}(x, y, ...)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}.}

\item{...}{Additional arguments (unused).}

\item{e1}{A sparse_numeric object (left operand)}

\item{e2}{A sparse_numeric object (right operand)}
}
\value{
A sparse_numeric object representing x - y.
}
\description{
Subtract two sparse_numeric vectors
}
