% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_6.R
\name{sparse_sum,sparse_numeric-method}
\alias{sparse_sum,sparse_numeric-method}
\title{Sum of sparse_numeric object}
\usage{
\S4method{sparse_sum}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{sparse_numeric object}
}
\value{
numeric scalar
}
\description{
Computes the sum of stored values in a sparse_numeric object.
}
\examples{
x <- as(c(1,0,2,0,3), "sparse_numeric")
sparse_sum(x)
}
