% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_6.R
\name{standardize,sparse_numeric-method}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse_numeric object}
\usage{
\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{sparse_numeric object}
}
\value{
sparse_numeric object
}
\description{
Standardizes a sparse_numeric object to mean 0 and standard deviation 1.
}
\examples{
x <- as(c(1,0,1), "sparse_numeric")
standardize(x)
}
