% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse_numeric object}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{...}{Unused.}
}
\value{
A standardized sparse_numeric vector.
}
\description{
Standardizes the underlying full vector by subtracting the mean
and dividing by the standard deviation. Only non-zero standardized
values are stored.
}
