% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Calculate Mean of Sparse Vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{...}{Additional arguments passed to sum().}
}
\value{
A numeric value.
}
\description{
Calculates the arithmetic mean of the sparse vector, including zeros.
}
