% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Mean of a sparse_numeric vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Unused.}
}
\value{
A numeric scalar.
}
\description{
Computes the mean of the underlying dense numeric vector, treating
unspecified entries as zero.
}
