% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot Overlapping Elements}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Additional graphical parameters.}
}
\description{
Plots the overlapping non-zero elements of two sparse_numeric vectors.
X-axis represents position, Y-axis represents value.
}
