% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\title{Sparse Addition}
\usage{
sparse_add(x, y, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y)

\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{a sparse_numeric object}

\item{y}{a sparse_numeric object}

\item{...}{additional arguments}

\item{e1}{a sparse numeric object}

\item{e2}{a sparse numeric object}
}
\value{
A new sparse_numeric object that represents the sum of x and y
}
\description{
Adds two sparse_numeric vectors
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_add(x = sparse_numeric, y = sparse_numeric)}: Add two sparse_numeric vectors

}}
