% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Sparse Cross Product}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Additional arguments.}
}
\value{
A single numeric value.
}
\description{
Calculates the dot product of two sparse_numeric vectors.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_crossprod(x = sparse_numeric, y = sparse_numeric)}: Calculate dot product

}}
