% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Sparse Multiplication}
\usage{
sparse_mult(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Additional arguments.}

\item{e1}{a sparse numeric object}

\item{e2}{a sparse numeric object}
}
\value{
A new sparse_numeric object representing the product.
}
\description{
Multiplies two sparse_numeric vectors element-wise.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_mult(x = sparse_numeric, y = sparse_numeric)}: Multiply two sparse_numeric vectors

}}
