% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_6.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse numeric vector class}
\description{
Stores a numeric vector in sparse format.
Only non-zero elements are stored.
}
\section{Slots}{

\describe{
\item{\code{value}}{numeric vector of non-zero values}

\item{\code{pos}}{integer vector of positions corresponding to values}

\item{\code{length}}{total length of the vector (including zeros)}
}}

