% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse Vector class}
\description{
An S4 class that represents sparse numeric vectors, or vectors which contain mostly zeros
}
\section{Slots}{

\describe{
\item{\code{value}}{a numeric vector containing the non-zero elements of the vector}

\item{\code{pos}}{an integer vector with the positions of the non-zero elements}

\item{\code{length}}{a single integer that indicates the full length of the vector}
}}

