% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\alias{sparse_numeric-class-validity}
\title{sparse_numeric S4 class}
\description{
sparse_numeric S4 class
}
\section{Slots}{

\describe{
\item{\code{value}}{Numeric vector of non-zero values.}

\item{\code{pos}}{Integer vector of positions (1-based) corresponding to each stored value.}

\item{\code{length}}{Integer scalar giving the total length of the full vector.}
}}

\keyword{internal}
