% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevector.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\title{Sparse Subtraction}
\usage{
sparse_sub(x, y, ...)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Additional arguments.}

\item{e1}{a sparse numeric object}

\item{e2}{a sparse numeric object}
}
\value{
A new sparse_numeric object representing the difference.
}
\description{
Subtracts one sparse_numeric vector from another.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_sub(x = sparse_numeric, y = sparse_numeric)}: Subtract two sparse_numeric vectors

}}
