% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize Sparse Vector}
\usage{
standardize(x)

\S4method{standardize}{sparse_numeric}(x)
}
\arguments{
\item{x}{A sparse_numeric object.}
}
\value{
A dense numeric vector where each element is (x_i - mean) / sd.
}
\description{
Standardizes the vector (subtract mean, divide by SD).
Returns a dense numeric vector because standardization usually destroys sparsity.
}
\section{Methods (by class)}{
\itemize{
\item \code{standardize(sparse_numeric)}: Standardize a sparse_numeric vector

}}
